Revert "Update Roboto font to 3.005"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3909d0c2-0969-4256-be97-c6ab0b614ae7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I974a8c5630fddb28e14c7edffbd5a53bb69d960d