Revert "Update Roboto font to 3.005"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5e56a409-5471-4377-86b9-e9bbc90fad7e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8ffe0a81f7f456e261de2ba73e28750cb70bd663