Revert "Update Roboto font to 3.005"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5f6b724d-4fef-447b-bf50-f42b2a0f007f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I14718437793e9e6bbf1ae8d64f9aeccf0209331e