Revert "Update Roboto font to 3.005"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/49ddd985-778a-4e55-9f7c-cdef8b3acefd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I56dba5d475cc232cb064e811245340a3d0a819c0