Revert "Update Roboto font to 3.005"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a55939f9-b31c-486a-9c5a-7818861b88fc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7650d7804c94652fa0007e09b611f3213bcde28c